Add files via upload
This was a simple assignment for a computer architecture course asking to find the sum of the first 100000 integers starting at zero, once with a single thread, and one with multi threading enabled. I chose to use 2 threads, but could have implemented up to 4. The FileIO allowed me to create a basic time sheet to keep track of execution times. The multi-threaded application had noticeably lower execution times in the trials.